+++ eruption/src/scripting/script.rs
@@ -519,10 +519,10 @@ fn realize_color_map() -> Result<RunningScriptResult> {
 
                     #[rustfmt::skip]
                     let color = RGBA {
-                        r: ((((fg.a as f32) * fg.r as f32 + (255 - fg.a) as f32 * bg.r as f32).floor() * brightness as f32 / 100.0) as u32 >> 8) as u8,
-                        g: ((((fg.a as f32) * fg.g as f32 + (255 - fg.a) as f32 * bg.g as f32).floor() * brightness as f32 / 100.0) as u32 >> 8) as u8,
-                        b: ((((fg.a as f32) * fg.b as f32 + (255 - fg.a) as f32 * bg.b as f32).floor() * brightness as f32 / 100.0) as u32 >> 8) as u8,
-                        a: fg.a as u8,
+                        r: ((((fg.a as f32) * fg.r as f32 + (255 - fg.a) as f32 * bg.r as f32).floor() * brightness / 100.0) as u32 >> 8) as u8,
+                        g: ((((fg.a as f32) * fg.g as f32 + (255 - fg.a) as f32 * bg.g as f32).floor() * brightness / 100.0) as u32 >> 8) as u8,
+                        b: ((((fg.a as f32) * fg.b as f32 + (255 - fg.a) as f32 * bg.b as f32).floor() * brightness / 100.0) as u32 >> 8) as u8,
+                        a: fg.a,
                     };
 
                     *background = color;
